use layout component
